encode: simplify a bunch of encoding implementations
The change from bitcoin 0.31 to 0.32 breaks the encoding API in some
pretty annoying ways, by moving from std::io to the nostd bitcoin::io
which has a weaker API due to Rust language limitations.

To get ahead of this, replace a bunch of manual little-endian code with
calls to {to,from}_le_bytes, and replace a bunch of passthroughs to
Bitcoin encoding code by a macro.